refactor(playground): base ApiDOM Playground on swagger-adjust
Closes #335
