+++ experiments/apidom-playground/package.json
@@ -20,7 +20,6 @@
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@reduxjs/toolkit": "^1.5.0",
     "classnames": "^2.2.6",
     "comlink": "^4.3.0",
     "install": "^0.13.0",
@@ -29,7 +28,8 @@
     "ramda-adjunct": "^2.30.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2"
+    "react-redux": "^7.2.2",
+    "swagger-adjust": "^0.3.0"
   },
   "devDependencies": {
     "@babel/core": "7.12.3",
