+++ experiments/apidom-playground/src/index.jsx
@@ -2,22 +2,47 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { CssBaseline } from '@material-ui/core';
+import * as Complink from 'comlink';
+import System, { SystemContext, useSystemComponent } from 'swagger-adjust';
 
 import './index.scss';
-import App from './features/app/components/App';
+import ApiDOMPlaygroundPlugin from './playground/plugin';
+import ApiDOMWorker from './playground/apidom.worker';
 import reportWebVitals from './reportWebVitals';
-import createStore from './store';
 
-const store = createStore();
+const system = new System({
+  plugins: [ApiDOMPlaygroundPlugin],
+  middleware: (sys) => (getDefaultMiddleware) => {
+    const apiDOMWorker = new ApiDOMWorker();
+    const apiDOMService = Complink.wrap(apiDOMWorker);
+
+    return getDefaultMiddleware({
+      thunk: {
+        extraArgument: { getSystem: sys.getSystem, apiDOMService },
+      },
+      serializableCheck: {
+        // ignore these field paths in all actions
+        ignoredActionPaths: ['payload'],
+      },
+    });
+  },
+});
+const store = system.getStore();
+
+const App = () => {
+  const Playground = useSystemComponent('Playground');
+
+  return <Playground />;
+};
 
 ReactDOM.render(
   <React.StrictMode>
-    <>
+    <SystemContext.Provider value={system.getSystem}>
       <CssBaseline />
       <Provider store={store}>
         <App />
       </Provider>
-    </>
+    </SystemContext.Provider>
   </React.StrictMode>,
   document.getElementById('root')
 );
